Node 기능 추가
아이템 적용 기능을
Node에 추가함
월드 아이템은 아직 안함

+++ SANCTUM_Online/SANCTUM_Server/Server/Game/Object/Enemy.cs
@@ -141,7 +141,7 @@ namespace Server.Game
             Room.Broadcast(lookPacket);
 
             // 피격
-            _target.OnDamaged(Owner, Stat.Attack);
+            _target.OnDamaged(Owner, (int)Stat.Attack);
         }
 
         protected virtual void UpdateDie()
